Use versioned base image with the runtime builder. (#296)
* Passing individual base image names to gen-dockerfile
* Dump the exact PHP versions in php-base image
* Use the new php56 image as the base of the test-runner
* Added test cases for using non-default PHP versions

+++ php-base/composer.sh
@@ -21,39 +21,42 @@ set -xe
 DEFAULT_PHP_VERSION="7.1"
 
 if [ -f ${APP_DIR}/composer.json ]; then
-    # Extract php version from the composer.json.
-    CMD="php /tmp/detect_php_version.php ${APP_DIR}/composer.json"
-    PHP_VERSION=`su www-data -c "${CMD}"`
+    if [ -n "${DETECTED_PHP_VERSION}" ]; then
+        PHP_VERSION="${DETECTED_PHP_VERSION}"
+    else
+        # Extract php version from the composer.json.
+        CMD="php /tmp/detect_php_version.php ${APP_DIR}/composer.json"
+        PHP_VERSION=`su www-data -c "${CMD}"`
 
-    # Remove files and directories for detecting PHP version.
-    # These files are created in Dockerfile.
-    rm -rf /tmp/vendor /tmp/detect_php_version.php /tmp/composer.*
+        # Remove files and directories for detecting PHP version.
+        # These files are created in Dockerfile.
+        rm -rf /tmp/vendor /tmp/detect_php_version.php /tmp/composer.*
 
-    if [ "${PHP_VERSION}" != "5.6" ] && [ "${PHP_VERSION}" != "7.0" ] && [ "${PHP_VERSION}" != "7.1" ]; then
-        cat<<EOF
+        if [ "${PHP_VERSION}" != "5.6" ] && [ "${PHP_VERSION}" != "7.0" ] && [ "${PHP_VERSION}" != "7.1" ]; then
+            cat<<EOF
 There is no PHP runtime version specified in composer.json, or we don't support the version you specified. Google App Engine uses the latest 7.1.x version. We recommend pinning your PHP version by running:
 
 composer require php 7.1.* (replace it with your desired minor version)
 
 Using PHP version 7.1.x...
 EOF
-        PHP_VERSION=${DEFAULT_PHP_VERSION}
-    fi
+            PHP_VERSION=${DEFAULT_PHP_VERSION}
+        fi
 
-    if [ "${PHP_VERSION}" == "5.6" ]; then
-        apt-get -y update
-        /bin/bash /build-scripts/install_php56.sh
-        apt-get remove -y gcp-php71
-        rm -rf /var/lib/apt/lists/*
-    fi
+        if [ "${PHP_VERSION}" == "5.6" ]; then
+            apt-get -y update
+            /bin/bash /build-scripts/install_php56.sh
+            apt-get remove -y gcp-php71
+            rm -rf /var/lib/apt/lists/*
+        fi
 
-    if [ "${PHP_VERSION}" == "7.0" ]; then
-        apt-get -y update
-        /bin/bash /build-scripts/install_php70.sh
-        apt-get remove -y gcp-php71
-        rm -rf /var/lib/apt/lists/*
+        if [ "${PHP_VERSION}" == "7.0" ]; then
+            apt-get -y update
+            /bin/bash /build-scripts/install_php70.sh
+            apt-get remove -y gcp-php71
+            rm -rf /var/lib/apt/lists/*
+        fi
     fi
-
     # Handle custom oauth keys (Adapted from https://github.com/heroku/heroku-buildpack-php/blob/master/bin/compile)
     COMPOSER_GITHUB_OAUTH_TOKEN=${COMPOSER_GITHUB_OAUTH_TOKEN:-}
     if [[ -n "$COMPOSER_GITHUB_OAUTH_TOKEN" ]]; then
